Update utils.py
Please explain the error:

RuntimeError: indices should be either on cpu or on the same device as the indexed tensor (cpu)
Gemini
Hypothesis
The error message "RuntimeError: indices should be either on cpu or on the same device as the indexed tensor (cpu)" indicates that you are trying to index a tensor that is on the CPU with indices that are on the GPU.

Looking at the traceback, the error occurs in utils.py within the calculate_mAP function, specifically at this line:

 
original_ind = torch.LongTensor(range(true_class_boxes.size(0)))[true_class_images == this_image][ind]
Use code with caution
Here's a breakdown of what's likely happening:

true_class_boxes is a tensor that resides on the CPU.
true_class_images and ind are tensors that likely reside on the GPU (since you are using device = torch.device("cuda" if torch.cuda.is_available() else "cpu") and loading your data to the device in the evaluate function).
You are trying to use true_class_images == this_image to create a boolean mask and then index into torch.LongTensor(range(true_class_boxes.size(0))) with this mask and ind.
Since the tensors are on different devices (CPU vs GPU), this operation fails.
Suggested Changes
To fix this error, you need to ensure that all tensors involved in the indexing operation are on the same device. Since you are primarily using the GPU for your model, it's best to move true_class_boxes to the GPU as well.

You will need to modify the calculate_mAP function in utils.py. You can likely add .to(device) to the end of the line generating true_class_boxes within the calculate_mAP function, so that the line now appears as this:

 
true_class_boxes = torch.FloatTensor(true_class_boxes).to(device)
Use code with caution
Here's the full calculate_mAP function with the change. Note that this assumes device is defined within or passed to the function calculate_mAP :

 
def calculate_mAP(det_boxes, det_labels, det_scores, true_boxes, true_labels, true_difficulties):
    """
    Calculate the Mean Average Precision (mAP) of detected objects.

    See https://medium.com/@jonathan_hui/map-mean-average-precision-for-object-detection-45c121a31173 for an explanation

    :param det_boxes: list of tensors, one tensor for each image containing detected objects' bounding boxes
    :param det_labels: list of tensors, one tensor for each image containing detected objects' labels
    :param det_scores: list of tensors, one tensor for each image containing detected objects' labels' scores
    :param true_boxes: list of tensors, one tensor for each image containing actual objects' bounding boxes
    :param true_labels: list of tensors, one tensor for each image containing actual objects' labels
    :param true_difficulties: list of tensors, one tensor for each image containing actual objects' difficulty (0 or 1)
    :return: list of average precisions for all classes, mean average precision (mAP)
    """
    assert len(det_boxes) == len(det_labels) == len(det_scores) == len(true_boxes) == len(
        true_labels) == len(
        true_difficulties)  # these are all lists of tensors of the same length, i.e. number of images
    n_classes = len(label_map)

    # Store all (true) objects for easier access
    true_class_images = list()
    for i in range(n_classes):
        true_class_images.append(list())
    for i in range(len(true_labels)):
        for j in range(true_labels[i].size(0)):
            true_class_images[true_labels[i][j]].append(i)
    true_class_images = [torch.LongTensor(x) for x in true_class_images]

    # Store all detections (boxes, labels, scores) for easier access
    true_class_boxes = list()
    true_class_labels = list()
    true_class_difficulties = list()
    true_class_scores = list()
    for i in range(n_classes):
        true_class_boxes.append(list())
Use code with caution
Sources
Rate this answer